Factor literals from underlying alphabet out into basic-symbols.ts for reuse, move alphabets for specific layers into their own folder
